Provide treeshakable animation frame api
In this diff I moved `cancel` from static property to export which will
allow webpack to treeshake it out when it is not unused.

To prevent creating `default` field in cjs I converted api to pair of
named exports (`raf` and `caf`) and renamed module to `animationFrame`.

Here's an output
```
var raf = exports.raf = function raf(cb) {
  return rafImpl(cb);
};
var caf = exports.caf = function caf(id) {
  window[cancel] && typeof window[cancel] === 'function' && window[cancel](id);
};
```

This is a breaking change and I think it should be considered
separately from treeshaking branch.